+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **uplofile**: Allow `Dropzone` `className` to be a function that receives drag state for styling.
 - **uplofile**: Enhance `beforeUpload` hook to accept previous items and remaining count.
 
 ## [2.2.2] - 2026-02-09